Move createEmotionCache to utils
