+++ examples/nextjs-with-typescript/src/pages/_app.tsx
@@ -5,7 +5,7 @@ import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider, EmotionCache } from '@emotion/react';
 import theme from '../../styles/theme';
-import createEmotionCache from '../lib/createEmotionCache';
+import createEmotionCache from '../utils/createEmotionCache';
 
 // Client-side cache, shared for the whole session of the user in the browser.
 const clientSideEmotionCache = createEmotionCache();
